added trip event item presenter
